Added Database Scripts
The create database script works fine, the crud one is fine too, but the other one may not run properly. I will have to update these manually from the google doc, since people are not using the GitHub. Ohwell. Also note that as of now there crud sp in the get_sp because I'm waiting on josh to fix them to update the db scripts.